move block device related files to 'block_device' folder
