+++ src/block_device/block_device_event.cpp
@@ -1,5 +1,5 @@
 #include "block_device_event.hpp"
-#include "exceptions.hpp"
+#include "../exceptions.hpp"
 
 BlockDeviceEventAction to_block_device_event_action(const std::string &action)
 {
