+++ src/block_device/block_device_event.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "block_device.hpp"
 #include "udev_event.hpp"
-#include "proto/block_device_event.pb.h"
+#include "../proto/block_device_event.pb.h"
 
 BlockDeviceEventAction to_block_device_event_action(const std::string &action);
